Fix PropertyLayer default value type mismatch warnings in tests
Updated PropertyLayer instantiations in tests/test_components_matplotlib.py and tests/test_space_renderer.py to explicitly use np.float64 for default values. This resolves UserWarnings about type mismatches between default values and the specified dtype.

+++ tests/test_components_matplotlib.py
@@ -2,6 +2,7 @@
 
 import matplotlib.pyplot as plt
 import networkx as nx
+import numpy as np
 
 from mesa import Agent, Model
 from mesa.discrete_space import (
@@ -222,14 +223,16 @@ def test_draw_property_layers():
     """Test drawing property layers."""
     model = Model(seed=42)
     grid = SingleGrid(10, 10, torus=True)
-    grid.add_property_layer(PropertyLayer("test", grid.width, grid.height, 0))
+    grid.add_property_layer(
+        PropertyLayer("test", grid.width, grid.height, np.float64(0))
+    )
 
     _, ax = plt.subplots()
     draw_property_layers(grid, {"test": {"colormap": "viridis", "colorbar": True}}, ax)
 
     model = Model(seed=42)
     grid = OrthogonalMooreGrid((10, 10), torus=True, random=model.random, capacity=1)
-    grid.create_property_layer("test", 0.0)
+    grid.create_property_layer("test", np.float64(0.0))
 
     _, ax = plt.subplots()
     draw_property_layers(grid, {"test": {"colormap": "viridis", "colorbar": True}}, ax)
